core: Refactor uv bindings
I can already see these are going to get massive. Putting them into multiple
files.
